refactor: restructure controllers by responsibility following SOLID principles
