+++ server/src/presentation/http/routes/chatRoutes.ts
@@ -2,22 +2,17 @@ import express from 'express';
 const router = express.Router();
 
 import {
-    sendMessageController,
-    getHistoryController,
     authenticatePsychologist,
     authenticateUser,
-    getUnreadCountController,
-    markAsReadController,
     authenticateAll,
-    getUserChatConsultationsController,
-    getPsychologistChatConsultationsController,
+    chatMessageController,
 } from '@/infrastructure/config/di';
 
-router.get('/patient', authenticateUser, getUserChatConsultationsController.handle);
-router.get('/psychologist', authenticatePsychologist, getPsychologistChatConsultationsController.handle);
-router.post('/send', sendMessageController.handle);
-router.get('/history/:consultationId', getHistoryController.handle);
-router.get('/:consultationId/unread-count', authenticateAll, getUnreadCountController.handle);
-router.put('/:consultationId/mark-read', authenticateAll, markAsReadController.handle);
+router.get('/patient', authenticateUser, chatMessageController.getUserChatConsultations);
+router.get('/psychologist', authenticatePsychologist, chatMessageController.getPsychologistChatConsultations);
+router.post('/send', chatMessageController.send);
+router.get('/history/:consultationId', chatMessageController.getHistory);
+router.get('/:consultationId/unread-count', authenticateAll, chatMessageController.getUnreadCount);
+router.put('/:consultationId/mark-read', authenticateAll, chatMessageController.markAsRead);
 
 export default router;
