+++ server/src/presentation/http/routes/psychologistRoutes.ts
@@ -1,27 +1,23 @@
 import express from 'express';
 import {
     authenticatePsychologist,
-    createSlotController,
-    deleteSlotController,
-    getSlotByPsychologistController,
-    updatePsychologistProfileController,
-    getKycDetailsController,
-    getProfileController,
     checkBlockedUser,
-    getPsychologistConsultationsController,
-    psychologistCancelConsultationController,
+    slotController,
+    psychologistConsultationController,
+    psychologistProfileController,
+    psychologistKycController,
 } from '@/infrastructure/config/di';
 
 import { uploadSingle } from '../middlewares/multer';
 
 const router = express.Router();
 
-router.get('/profile', authenticatePsychologist, getProfileController.handle);
-router.put('/profile',authenticatePsychologist, uploadSingle, updatePsychologistProfileController.handle);
-router.get('/kyc', authenticatePsychologist, getKycDetailsController.handle);
-router.post('/slot', authenticatePsychologist, createSlotController.handle);
-router.get('/slot', authenticatePsychologist, getSlotByPsychologistController.handle);
-router.delete('/slot/:slotId', authenticatePsychologist, deleteSlotController.handle);
-router.get('/consultations', authenticatePsychologist, checkBlockedUser, getPsychologistConsultationsController.handle);
-router.put('/consultation/:id/cancel', authenticatePsychologist, checkBlockedUser, psychologistCancelConsultationController.handle);
+router.get('/profile', authenticatePsychologist, psychologistProfileController.getProfile);
+router.put('/profile',authenticatePsychologist, uploadSingle, psychologistProfileController.updateProfile);
+router.get('/kyc', authenticatePsychologist, psychologistKycController.getKycDetails);
+router.post('/slot', authenticatePsychologist, slotController.createSlot);
+router.get('/slot', authenticatePsychologist, slotController.getSlotByPsychologist);
+router.delete('/slot/:slotId', authenticatePsychologist, slotController.deleteSlot);
+router.get('/consultations', authenticatePsychologist, checkBlockedUser, psychologistConsultationController.getConsultations);
+router.put('/consultation/:id/cancel', authenticatePsychologist, checkBlockedUser, psychologistConsultationController.cancelConsultation);
 export default router;
